add new file : list_parsing.c to keep chained list function for arguments
