+++ .history_cmd
@@ -251,3 +251,7 @@ mama | papa | saucisse
 mama | papa | saucisse
 mama | papa | sauccisse
 mama | papa | saucisse
+mama | saucisse | papa
+mama | papa | saucisse
+mama | sauccisee | papa
+mama | papa | saucisse
